+++ lib/eva/lurp.py
@@ -57,10 +57,33 @@ def stop():
 
 
 def _t_dispatcher(host, port):
+    chunk_buf = {}
     server_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     server_socket.bind((host, port))
     logging.debug('LURP dispatcher started')
     controller = eva.core.controllers[0]
+
+    def process_message(msg):
+        if msg['s'] == 'state':
+            state = msg['d']
+            for s in [state] if isinstance(state, dict) else state:
+                item = controller.get_item(s['oid'])
+                if item:
+                    if s['type'] == 'unit':
+                        result = item.set_state_from_serialized(s, notify=False)
+                        if result:
+                            need_notify = item.update_nstate(
+                                nstatus=s['nstatus'], nvalue=s['nvalue'])
+                            if item.action_enabled != s['action_enabled']:
+                                item.action_enabled = s['action_enabled']
+                                need_notify = True
+                            if result == 2 or need_notify:
+                                item.notify()
+                    else:
+                        item.set_state_from_serialized(s)
+                else:
+                    logging.debug(f'LURP item not found: {s["oid"]}')
+
     while _flags.dispatcher_active:
         try:
             data, address = server_socket.recvfrom(config.buf)
@@ -68,36 +91,25 @@ def _t_dispatcher(host, port):
                 return
             if not data:
                 continue
-            address = address[0]
-            logging.debug('LURP message from %s' % address)
+            logging.debug('LURP message from %s' % address[0])
             try:
-                msg = msgpack.loads(data[2:], raw=False)
-                if msg['s'] == 'state':
-                    state = msg['d']
-                    for s in [state] if isinstance(state, dict) else state:
-                        item = controller.get_item(s['oid'])
-                        if item:
-                            if s['type'] == 'unit':
-                                result = item.set_state_from_serialized(
-                                    s, notify=False)
-                                if result:
-                                    need_notify = item.update_nstate(
-                                        nstatus=s['nstatus'],
-                                        nvalue=s['nvalue'])
-                                    if item.action_enabled != s[
-                                            'action_enabled']:
-                                        item.action_enabled = s[
-                                            'action_enabled']
-                                        need_notify = True
-                                    if result == 2 or need_notify:
-                                        item.notify()
-                            else:
-                                item.set_state_from_serialized(s)
-                        else:
-                            logging.debug(f'LURP item not found: {s["oid"]}')
+                if data[0] == 0x01:  # first proto version
+                    msg = msgpack.loads(data[2:], raw=False)
+                    process_message(msg)
+                elif data[2] == 0x00:  # single frame received
+                    msg = msgpack.loads(data[3:], raw=False)
+                    process_message(msg)
+                elif data[2] == 0x01:  # first chunk received
+                    chunk_buf[address] = data[3:]
+                elif data[2] == 0x02:  # next chunk received
+                    chunk_buf[address] += data[3:]
+                elif data[2] == 0xff:  # last chunk received
+                    frame = chunk_buf.pop(address) + data[3:]
+                    msg = msgpack.loads(frame, raw=False)
+                    process_message(msg)
             except Exception as e:
                 logging.error(
-                    f'LURP message from {address} processing failed: {e}')
+                    f'LURP message from {address[0]} processing failed: {e}')
                 eva.core.log_traceback()
         except:
             logging.critical('LURP dispatcher crashed, restarting')